Properly implements data serialization and deserialization for GlobalResearchTree, only sent when first connecting to a physical server, or when datapacks on a physical server are reloaded.  Apologies ahead of time for what I did to GlobalResearchFactory.
Adds the "branch-id" and "base-time" tags, which can be used to set the (highly estimated!) expected time for a given research branch.  These files can also contain "name" tags which apply to the branch label text as translation text.  Removes the configuration setting to change research speed.

Adds the "subtitle" tag to Researches and Research Effects.  This does need to be a fully-qualified translation key or the string intended to present to users, rather than auto-generated from id, since it's an optional tag.  Only a few items contain it for now.

Removes research-related colony ticking.  Add colony research building requirement checking events to AbstractEntityAIStructureWithWorkOrder's executeSpecificCompleteActions and moved the logic to handle such into ResearchManager. Moves as much of the logic as is practical from TryResearchMessage to LocalResearchTree.

Adds sound effects to some research events and failures.

Switches ItemUtils.tryRemoveStackFromItemHandler and tryRemoveStorageFromItemHandler to check for item presence then remove, rather than previous spool/unspool-on-fail.

Lots of code cleanup: switches to doubles whenever floats were used, gets rid of unused imports, properly wildcards generics when correct, moves some Lists to HashSets, removes duplicates from building registry and adjusts properties to match.  Hopefully managed to keep newlines in line.  Fixes a lot of bad json formatting.

+++ src/api/java/com/minecolonies/api/research/effects/IResearchEffectManager.java
@@ -24,7 +24,7 @@ public interface IResearchEffectManager
      * @param id   the id of the effect.
      * @return the strength of the effect, or zero if it isn't present.
      */
-     float getEffectValue(final String id);
+     double getEffectValue(final String id);
 
     /**
      * Gets whether a research effect is enabled with a non-zero value.
